Move some functionality to a dedicated ROS2 node interface for reusability
More ROS2 nodes will be created that depend on a valid RON node ID, so move this
to its own library. This calls for the creation of the ron_common package, which
contains utilities usable by many RON packages.

While we're at it, fix methods being moved to this base class because they had
improper case.